src/requirements.txt
duckdb==1.3.1
mysql-connector-python==9.3.0
numpy==2.2.6
openai==1.88.0
openpyxl==3.1.5
pandas==2.3.0
pydantic==2.11.7
pymongo==4.13.2
PyMySQL==1.1.1
psycopg2==2.9.9
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
SQLAlchemy==2.0.41
tabulate==0.9.0
tqdm==4.67.1
httpx==0.28.1
